refactor and lazy in progress again
