+++ grakn-dashboard/src/components/graphPage/graphPage.vue
@@ -43,14 +43,9 @@ along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
 
 <script>
 // Modules
-import HALParser, {
-    URL_REGEX,
-} from '../../js/HAL/HALParser';
-import * as API from '../../js/util/HALTerms';
-import EngineClient from '../../js/EngineClient';
-import User from '../../js/User';
-import Visualiser from '../../js/visualiser/Visualiser';
+import HALParser from '../../js/HAL/HALParser';
 import GraphPageState from '../../js/state/graphPageState';
+import CanvasHandler from './modules/CanvasHandler';
 
 // Sub-components
 const NodePanel = require('./nodePanel.vue');
@@ -69,9 +64,7 @@ export default {
     data() {
         return {
             state: GraphPageState,
-            halParser: {},
-            analyticsStringResponse: undefined,
-            doubleClickTime: 0,
+            canvasHandler: {},
             selectedNodeLabel: undefined,
             codeMirror: {},
             allNodeOntologyProps: {},
@@ -86,27 +79,18 @@ export default {
     },
 
     created() {
-        window.visualiser = new Visualiser();
-
-        visualiser.setCallbackOnEvent('click', this.singleClick);
-        visualiser.setCallbackOnEvent('doubleClick', this.doubleClick);
-        visualiser.setCallbackOnEvent('oncontext', this.rightClick);
-        visualiser.setCallbackOnEvent('hold', this.holdOnNode);
-        visualiser.setCallbackOnEvent('hoverNode', this.hoverNode);
-        visualiser.setCallbackOnEvent('blurNode', this.blurNode);
-        visualiser.setCallbackOnEvent('dragStart', this.onDragStart);
-        visualiser.setCallbackOnEvent('stabilized', this.onStabilized);
-
-        this.halParser = new HALParser();
-
-        this.halParser.setNewResource((id, p, a, l, cn) => visualiser.addNode(id, p, a, l, cn));
-        this.halParser.setNewRelationship((f, t, l) => visualiser.addEdge(f, t, l));
-        this.halParser.setNodeAlreadyInGraph(id => visualiser.nodeExists(id));
+        this.canvasHandler = new CanvasHandler(this.state);
 
         // Register listened on State events
-        this.state.eventHub.$on('click-submit', this.onClickSubmit);
-        this.state.eventHub.$on('load-ontology', this.onLoadOntology);
+        this.state.eventHub.$on('click-submit', query=>this.canvasHandler.onClickSubmit(query));
+        this.state.eventHub.$on('load-ontology', type=>this.canvasHandler.onLoadOntology(type));
         this.state.eventHub.$on('clear-page', this.onClear);
+        //Events from canvasHandler
+        this.state.eventHub.$on('show-node-panel', this.onShowNodePanel);
+        this.state.eventHub.$on('hover-node', this.onHoverNode);
+        this.state.eventHub.$on('blur-node', this.onBlurNode);
+
+
     },
     beforeDestroy() {
         // Destroy listeners when component is destroyed - although it never gets detroyed for now. [keep-alive]
@@ -121,259 +105,16 @@ export default {
             // TODO: find a way to compute this without jQuery:
             this.graphOffsetTop = $('#graph-div').offset().top;
 
-            visualiser.render(graph);
+            this.canvasHandler.renderGraph(graph);
         });
     },
 
     methods: {
-
-        onStabilized(){
-          console.log("stabilizzato");
-          this.halParser.instances.forEach((uri) => {
-            EngineClient.request({
-              url: uri,
-            }).then((resp) => {
-              const respObj = JSON.parse(resp);
-              visualiser.updateNodeResources(respObj[API.KEY_ID], Utils.extractResources(respObj));
-            }, (err) => {
-              console.log(err);
-            });
-          });
-        },
-
-        updateRectangle(e) {
-            visualiser.updateRectangle(e.pageX, e.pageY - this.graphOffsetTop);
-        },
-
-        onLoadOntology(type) {
-            const querySub = `match $x sub ${type};`;
-            EngineClient.graqlHAL(querySub).then(this.onGraphResponse, (err) => {
-                this.state.eventHub.$emit('error-message', err.message);
-            });
-        },
-        onClickSubmit(query) {
-            if (query.includes('aggregate')) {
-                // Error message until we will not properly support aggregate queries in graph page.
-                this.state.eventHub.$emit('error-message', 'Invalid query: \'aggregate\' queries are not allowed from the Graph page. Please use the Console page.');
-                return;
-            }
-
-            if (query.trim().startsWith('compute')) {
-                EngineClient.graqlAnalytics(query).then(this.onGraphResponseAnalytics, (err) => {
-                    this.state.eventHub.$emit('error-message', err.message);
-                });
-            } else {
-                let queryToExecute = query.trim();
-
-                if (!(query.includes('offset')) && !(query.includes('delete')))
-                    queryToExecute = queryToExecute + ' offset 0;';
-                if (!(query.includes('limit')) && !(query.includes('delete')))
-                    queryToExecute = queryToExecute + ' limit ' + User.getQueryLimit() + ';';
-                this.emitInjectQuery(queryToExecute);
-                EngineClient.graqlHAL(queryToExecute).then(this.onGraphResponse, (err) => {
-                    this.state.eventHub.$emit('error-message', err.message);
-                });
-            }
-        },
-
-        checkSelectionRectangleStatus(node, eventKeys, param) {
-            // If we were drawing rectangle and we click again we stop the drawing and compute selected nodes
-            if (visualiser.draggingRect) {
-                visualiser.draggingRect = false;
-                visualiser.resetRectangle();
-                visualiser.network.redraw();
-            } else {
-                if (eventKeys.ctrlKey && node === undefined) {
-                    visualiser.draggingRect = true;
-                    visualiser.startRectangle(param.pointer.canvas.x, param.pointer.canvas.y - this.graphOffsetTop);
-                }
-            }
-
-        },
-
-        configureNode(nodeType, selectedProps) {
-            visualiser.setDisplayProperties(nodeType, selectedProps);
-        },
-        onGraphResponseAnalytics(resp) {
-          const responseObject = JSON.parse(resp);
-            if (responseObject.type === 'string') {
-                this.state.eventHub.$emit('analytics-string-response', responseObject.response);
-            } else {
-                this.halParser.parseResponse(responseObject.response, false, false);
-                visualiser.fitGraphToWindow();
-            }
-        },
-
-        onGraphResponse(resp, nodeId) {
-            const responseObject = JSON.parse(resp);
-            if (!this.halParser.parseResponse(responseObject, false, false, nodeId)) {
-                this.state.eventHub.$emit('warning-message', 'No results were found for your query.');
-            } else {
-                // When a nodeId is provided is because the user double-clicked on a node, so we need to update its href
-                // which will contain a new value for offset
-                if (nodeId) {
-                    visualiser.nodes.update({
-                        id: nodeId,
-                        href: responseObject._links.self.href,
-                    });
-                }
-            }
-            visualiser.fitGraphToWindow();
-        },
-
-        onGraphResponseOntology(resp, nodeId) {
-            if (!this.halParser.parseResponse(JSON.parse(resp), true, true, nodeId)) {
-                this.state.eventHub.$emit('warning-message', 'No results were found for your query.');
-            }
-            visualiser.fitGraphToWindow();
-        },
-
-        onClear() {
-            // Reset all interface elements to default.
-            this.showNodeLabelPanel = false;
-            this.showNodePanel = false;
-
-            // And clear the graph
-            visualiser.clearGraph();
-        },
-        ////// Emits and page elements related methods  ///////
-
-        /**
-         * Prepare the list of resources to be shown in the right div panel
-         * It sorts them alphabetically and then check if a resource value is a URL
-         */
-        prepareResources(originalObject) {
-            if (originalObject == null) return {};
-            return Object.keys(originalObject).sort().reduce(
-                // sortedObject is the accumulator variable, i.e. new object with sorted keys
-                // k is the current key
-                (sortedObject, k) => {
-                    // Add 'href' field to the current object, it will be set to TRUE if it contains a valid URL, FALSE otherwise
-                    const currentResourceWithHref = Object.assign({}, originalObject[k], {
-                        href: this.validURL(originalObject[k].label)
-                    });
-                    return Object.assign({}, sortedObject, {
-                        [k]: currentResourceWithHref
-                    });
-                }, {});
-        },
-        validURL(str) {
-            const pattern = new RegExp(URL_REGEX, 'i');
-            return pattern.test(str);
-        },
-        emitInjectQuery(query) {
-            this.showContextMenu = false;
-            this.state.eventHub.$emit('inject-query', query);
-        },
-
-        ////////////////////////////////////////////////////// ----------- Graph mouse interactions ------------ ///////////////////////////////////////////////////////////
-
-        holdOnNode(param) {
-            const node = param.nodes[0];
-            if (node === undefined) return;
-
-            this.state.eventHub.$emit('show-label-panel', visualiser.getAllNodeProperties(node), visualiser.getNodeType(node), node);
-        },
-
-        doubleClick(param) {
-            this.doubleClickTime = new Date();
-            const node = param.nodes[0];
-            if (node === undefined) {
-                return;
-            }
-
-            const eventKeys = param.event.srcEvent;
-            const nodeObj = visualiser.getNode(node);
-
-            if (eventKeys.shiftKey) {
-                this.requestOntology(nodeObj);
-            } else {
-                let generatedNode = false;
-                //If we are popping a generated relationship we need to append the 'reasoner' parameter to the URL
-                if (nodeObj.baseType === API.GENERATED_RELATION_TYPE) {
-                    generatedNode = true;
-                }
-
-                EngineClient.request({
-                    url: nodeObj.href,
-                    appendReasonerParams: generatedNode,
-                }).then((resp) => this.onGraphResponse(resp, node), (err) => {
-                    this.state.eventHub.$emit('error-message', err.message);
-                });
-                if (generatedNode) {
-                    visualiser.deleteNode(node);
-                }
-            }
-        },
-        rightClick(param) {
-            const node = param.nodes[0];
-            if (node === undefined) {
-                return;
-            }
-
-            if (param.event.shiftKey) {
-                param.nodes.forEach((x) => {
-                    visualiser.deleteNode(x);
-                });
-            }
-        },
-
-        hoverNode(param) {
-            // Mouse event becomes position of hovered node
-            this.mouseEvent = param;
-            this.showToolTip = true;
-        },
-        blurNode() {
-            this.showToolTip = false;
-        },
-        requestOntology(nodeObj) {
-            // If alt key is pressed we load ontology related to the current node
-            if (nodeObj.ontology) {
-                EngineClient.request({
-                    url: nodeObj.ontology,
-                }).then((resp) => this.onGraphResponseOntology(resp, nodeObj.id), (err) => {
-                    this.state.eventHub.$emit('error-message', err.message);
-                });
-            }
-        },
-        leftClick(param) {
-
-            // TODO: handle multiselect properly now that is enabled.
-            const node = param.nodes[0];
-            const eventKeys = param.event.srcEvent;
-            const clickType = param.event.type;
-
-            // If it is a long press on node: return and onHold() method will handle the event.
-            if (clickType !== 'tap') {
-                return;
-            }
-
-            // Check if we need to start or stop drawing the selection rectangle
-            this.checkSelectionRectangleStatus(node, eventKeys, param);
-
-            if (node === undefined) {
-                return;
-            }
-
-            const nodeObj = visualiser.getNode(node);
-
-            if (eventKeys.shiftKey) {
-                this.requestOntology(nodeObj);
-            } else {
-
-                // Show node properties on node panel.
-                this.allNodeOntologyProps = {
-                    id: nodeObj.id,
-                    type: nodeObj.type,
-                    baseType: nodeObj.baseType,
-                };
-
-                this.allNodeResources = this.prepareResources(nodeObj.properties);
-                this.selectedNodeLabel = visualiser.getNodeLabel(node);
-
-                this.showNodePanel = true;
-            }
-
+        onShowNodePanel(ontologyProps,resources,label) {
+          this.allNodeOntologyProps = ontologyProps;
+          this.allNodeResources = resources;
+          this.selectedNodeLabel = label;
+          this.showNodePanel = true;
         },
 
         customContextMenu(e) {
@@ -384,42 +125,44 @@ export default {
             }
         },
 
-        singleClick(param) {
-            // Everytime the user clicks on canvas we clear the context-menu and tooltip
+        onHoverNode(param) {
+            // Mouse event becomes position of hovered node
+          this.mouseEvent = param;
+          this.showToolTip = true;
+        },
+        onBlurNode(){
+          this.showToolTip = false;
+        },
+
+        updateRectangle(e) {
+            visualiser.updateRectangle(e.pageX, e.pageY - this.graphOffsetTop);
+        },
+
+
+        configureNode(nodeType, selectedProps) {
+            visualiser.setDisplayProperties(nodeType, selectedProps);
+        },
+
+        onClear() {
+            // Reset all interface elements to default.
+            this.showNodeLabelPanel = false;
+            this.showNodePanel = false;
+
+            // And clear the graph
+            this.canvasHandler.clearGraph();
+        },
+        ////// Emits and page elements related methods  ///////
+
+
+        emitInjectQuery(query) {
             this.showContextMenu = false;
-            this.showToolTip = false;
-
-            const t0 = new Date();
-            const threshold = 200;
-            // all this fun to be able to distinguish a single click from a double click
-            if (t0 - this.doubleClickTime > threshold) {
-                setTimeout(() => {
-                    if (t0 - this.doubleClickTime > threshold) {
-                        this.leftClick(param);
-                    }
-                }, threshold);
-            }
+            this.state.eventHub.$emit('inject-query', query);
         },
 
-        onDragStart(params) {
-            const eventKeys = params.event.srcEvent;
-            visualiser.draggingNode = true;
-            this.showToolTip = false;
-            //If ctrl key is pressed while dragging node/nodes we also unlock and drag the connected nodes
-            if (eventKeys.ctrlKey) {
-                let neighbours = [];
-                params.nodes.forEach(node => {
-                    neighbours.push.apply(neighbours, visualiser.network.getConnectedNodes(node));
-                    neighbours.push(node);
-                });
-                visualiser.network.selectNodes(neighbours);
-                visualiser.releaseNodes(neighbours);
-            } else {
-                visualiser.releaseNodes(params.nodes);
-            }
-        },
+        onGraphResponse(resp){
+          this.canvasHandler.onGraphResponse(resp);
+        }
 
-        // ----- End of graph interactions ------- //
     },
 };
 </script>
